// labify-front-main/src/app/core/models/cart-product.model.ts
export type CartProduct = {
  id: 2;
  quantity: 3;
  variant: {
    id: 23;
    additional_price: 67;
    dimension_id: 30;
    color_id: 41;
    image: 'https://labify-website-backend.developnetwork.net/storage/uploads/Variant/651bc9592b6d8-20231003-075713.png';
    color: {
      id: 41;
      name: 'Chocolate';
      code: '#D2691E';
    };
    dimension: {
      id: 30;
      dimension: 'Eaque pariatur Est';
    };
    images: [
      'https://labify-website-backend.developnetwork.net/storage/uploads/Variant/651bc9592b6d8-20231003-075713.png'
    ];
  };
  product: {
    id: 20;
    name: 'A officia eos sint';
    sku: null;
    description: null;
    slug: 'a-officia-eos-sint-et-possimus-accusam';
    price: 46;
    image: 'https://labify-website-backend.developnetwork.net/storage/uploads/Variant/651bc9592b6d8-20231003-075713.png';
    currency: 'pound';
    availability: 'in stock';
    price_after_discount: 77;
    have_discount: false;
    variants: [
      {
        id: 23;
        additional_price: 67;
        dimension_id: 30;
        color_id: 41;
        image: 'https://labify-website-backend.developnetwork.net/storage/uploads/Variant/651bc9592b6d8-20231003-075713.png';
        color: {
          id: 41;
          name: 'Chocolate';
          code: '#D2691E';
        };
        dimension: {
          id: 30;
          dimension: 'Eaque pariatur Est';
        };
        images: [
          'https://labify-website-backend.developnetwork.net/storage/uploads/Variant/651bc9592b6d8-20231003-075713.png'
        ];
      }
    ];
    colors: [
      {
        id: 41;
        name: 'Chocolate';
        code: '#D2691E';
      }
    ];
    tag: {
      id: 5;
      name: 'Verdie Parker collection';
    };
    category: {
      id: 1;
      name: 'Lonzo Brown collection';
    };
  };
};